// src/app/shared/models/document.Model.ts
export class DocumentModel {
    DocumentId: string = '';
    LovId: string = '';
    Table: string = '';
    TableRefrenceId: string = '';
    DocumentPath: string = '';
    DocumentName: string = '';
    DocumentSize: string = '';
    DocumentExtention: string = '';
    DocumentGuid: string = '';
    Base64Url: string | ArrayBuffer | null = '';
  }